Split into subpackage + introduce LayerKVQuantState
Three structural cleanups, no behavior change (bit-identical to KIVI's
reference kernels post-split: max_diff=0.00e+00, torch.equal on packed
codes).

(1) Split the 520-line kv_fake_quant.py into a 4-file subpackage by
    concern:

      kv_fake_quant/
        __init__.py        re-exports + ASCII call-graph at the top
        kernels.py         FP8 + int4 custom_ops + math
        layer_hooks.py     LayerKVQuantState + attach + apply + calib cache
        fusion.py          smoothkv_fused weight-folding + maybe_run_*

    Reviewers can navigate by concern: "the math" -> kernels.py;
    "how it's wired into Attention" -> layer_hooks.py; "the load-time
    optimization" -> fusion.py.

(2) Introduce LayerKVQuantState, an nn.Module that owns all per-layer
    KV-quant state in one place: method (str), group_size (int),
    bits (int), and -- for SmoothKV -- s_k / s_v as non-persistent
    buffers (auto-migrate with module.to(device)).

    Replaces 5 loose attrs scattered across the layer instance:
      _kv_quant_method, _kv_quant_group_size, _kv_quant_bits,
      _kv_quant_s_k, _kv_quant_s_v   ->   layer.kv_quant_state

    Subclassing nn.Module (rather than dataclass) is what makes the
    SmoothKV calib buffers auto-migrate to GPU.

    Inline edit in attention.py (forward) updated to check
    `kv_quant_state` instead of `_kv_quant_method`.

(3) Move the misplaced `import re` from mid-file to the top of fusion.py
    (was at line 374 of the old kv_fake_quant.py, hard to spot).
